refactored, changing file names, using a baseURL in api.js, and moving file to utils folder
