--- src/api.js
@@ -1,5 +1,0 @@
-import axios from 'axios';
-
-export const fetchReviews = () => {
-    return axios("https://nc-games-2enb.onrender.com/api/reviews");
-}
