+++ src/utils/api.js
@@ -1,0 +1,17 @@
+import axios from 'axios';
+
+const reviews = axios.create({
+    baseURL: "https://nc-games-2enb.onrender.com/api/reviews"
+})
+
+
+export const fetchReviews = () => {
+    return reviews
+    .get(`/`)
+    .then(({ data }) => {
+        return data.reviews;
+    })
+    .catch((error) => {
+        console.log('error', error)
+    })
+}
